fix: corregir lógica de rutas y middlewares
