+++ usuarios/control/routes/user.routes.js
@@ -1,11 +1,17 @@
 const express = require('express');
-const usuariosController = require('../controllers/user.controller')
+const userController = require('../controllers/user.controller')
 
 const router = express.Router()
 
-router.get('/login', usuariosController.getLogin)
-router.get('/registrar', usuariosController.getSignUp)
-router.post('/login', usuariosController.postLogin)
-router.post('/registrar', usuariosController.postSignUp)
+router.get('/login', userController.getLogin)
+router.get('/registrar', userController.getSignUp)
+router.post('/login', userController.postLogin)
+router.post('/registrar', userController.postSignUp)
+
+
+// Permisos
+router.get('/', userController.getPermissions);
+
+router.get('/edit/:firebaseUID', userController.postEditPermissions);
 
 module.exports = router;
